[O] Reuse export interface
